# ⚡ CLI Discovery Performance Enhancement: From Zero to Sub-Millisecond Hero