--- a/Gemfile
+++ b/Gemfile
@@ -5,6 +5,9 @@ source 'https://rubygems.org'
 # Core CLI framework
 gem 'thor', '~> 1.3'
 
+# Concurrency and performance
+gem 'concurrent-ruby', '~> 1.2'
+
 # Development and testing dependencies
 group :development, :test do
   gem 'rspec', '~> 3.12'