--- a/Gemfile
+++ b/Gemfile
@@ -8,6 +8,9 @@ gem 'thor', '~> 1.3'
 # Concurrency and performance
 gem 'concurrent-ruby', '~> 1.2'
 
+# Compression for cache optimization
+gem 'lz4-ruby', '~> 0.3'
+
 # Development and testing dependencies
 group :development, :test do
   gem 'rspec', '~> 3.12'