--- a/TASK.md
+++ b/TASK.md
@@ -1,0 +1,20 @@
+
+## Priority 2: Discoverability (Know What's Available)
+
+### List Available Categories
+```bash
+leyline categories
+# Shows all categories with description and file count
+```
+
+### Show Category Contents
+```bash
+leyline show typescript
+# Lists all bindings in typescript category
+```
+
+### Search Standards
+```bash
+leyline search "testing"
+# Finds all tenets/bindings mentioning testing
+```