… BACKLOG.md
… Gemfile.lock
… TODO.md
… spec/lib/leyline/cli_backward_compatibility_spec.rb
… spec/lib/leyline/cli_discovery_spec.rb
… spec/lib/leyline/cli_warmup_integration_spec.rb
… spec/lib/leyline/discovery/cache_hit_ratio_validation_spec.rb
… spec/lib/leyline/discovery/document_scanner_parallel_spec.rb
… spec/lib/leyline/discovery/metadata_cache_compression_spec.rb
… spec/lib/leyline/discovery/metadata_cache_fuzzy_search_spec.rb
… spec/lib/leyline/discovery/metadata_cache_memory_accounting_spec.rb
… spec/lib/leyline/discovery/metadata_cache_performance_spec.rb
… spec/lib/leyline/discovery/metadata_cache_warmup_spec.rb
… spec/lib/leyline/discovery/performance_benchmark_spec.rb
… spec/lib/leyline/discovery/performance_validation_spec.rb
… spec/performance/discovery_performance_spec.rb